fix: suppress compiler warnings and fix type conversions
- Add pragma directives to suppress -Wsuggest-attribute=noreturn warning in graph_reserve()
- Fix float literal precision in llama_hparams (0.1 to 0.1f)
- Add static_cast<uint32_t> to fix signed/unsigned conversion warnings in KV cells
- Change tensor_idx type from int to int64_t to match gguf_find_tensor return type

+++ src/llama-kv-cells.h
@@ -57,7 +57,7 @@ public:
     }
 
     uint32_t size() const {
-        return pos.size();
+        return static_cast<uint32_t>(pos.size());
     }
 
     void resize(uint32_t n) {
@@ -77,7 +77,7 @@ public:
     }
 
     uint32_t get_used() const {
-        return used.size();
+        return static_cast<uint32_t>(used.size());
     }
 
     // the index of the first cell that is used
@@ -294,7 +294,7 @@ public:
         assert(i < pos.size());
         assert(pos[i] != -1);
 
-        return seq[i].count();
+        return static_cast<int>(seq[i].count());
     }
 
     // check if the cell contains seq_id
